// standart/src/global-types.ts
// Auto-generated global type definitions for ikas code components.
// This file is regenerated automatically — do not edit manually.

/** Enum type: CTMWlOfqfo */
export type BlogCardRatio = "1/1" | "4/3";
/** Enum type: 5Y6qMgnl0u */
export type ButtonStyle = "primary" | "secondary" | "ghost";
/** Enum type: scZQrZrXwk */
export type ContentAlignment = "left" | "center" | "right";
/** Enum type: iVUvUxinej */
export type ContentPosition = "top" | "center" | "bottom";
/** Enum type: WdAlLmxf65 */
export type FilterLayout = "sidebar" | "inline";
/** Enum type: E9QwAHCe3A */
export type FilterPosition = "left" | "right";
/** Enum type: YgE3J7LJvS */
export type ImageAspectRatio = "4/3" | "1/1" | "2/3";
/** Enum type: JG4LjKd1gk */
export type ImagePosition = "left" | "right";
/** Enum type: Q2vzViWEt0 */
export type ImageSlideObjectFit = "cover" | "contain";
/** Enum type: DOJOsE0Gkm */
export type LogoPosition = "left" | "center";
/** Enum type: iOemtHmkJv */
export type MediaLinkGridLayout = "bento-6" | "featured-3" | "grid-4" | "grid-6";
/** Enum type: JeKdY7zE5C */
export type MediaType = "image" | "video";
/** Enum type: yScKjz4g17 */
export type OverlayDirection = "bottom-up" | "center";
/** Enum type: oNss7rnNqX */
export type PaginationType = "load-more" | "numbered";
/** Enum type: MDdmh9Dz79 */
export type PdpGalleryPosition = "left" | "right";
/** Enum type: HSnRxNAQjK */
export type PdpGalleryView = "thumbnails" | "grid";
/** Enum type: A4PQWmePd0 */
export type PdpThumbnailPosition = "bottom" | "left";
/** Enum type: XQxBBItnl9 */
export type ProductListColumns = "2" | "3" | "4";
/** Enum type: 0sfpxAB4ZZ */
export type ProductListColumnsTablet = "1" | "2";
/** Enum type: WOFwVvW8ru */
export type ProductTitleSize = "large" | "small";
/** Enum type: ovl1tZFYJq */
export type SavingsBadgeStyle = "amount" | "percent";
/** Enum type: XEHFfzdUCt */
export type SectionHeight = "100vh" | "80vh" | "auto";
/** Enum type: itp4PRR11b */
export type SectionWidth = "contained" | "full-width";
/** Enum type: ULPXiTcmFl */
export type SliderAspectRatio = "3/4" | "1/1";